+++ Opcodes/locsig.c
@@ -45,7 +45,7 @@ static int locsigset(CSOUND *csound, LOCSIG *p)
     if (p->auxch.auxp == NULL ||
         p->auxch.size<sizeof(MYFLT)*(csound->ksmps * 4)) {
       MYFLT *fltp;
-      csound->AuxAlloc(csound, (long) (csound->ksmps * 4)
+      csound->AuxAlloc(csound, (size_t) (csound->ksmps * 4)
                                * sizeof(MYFLT), &p->auxch);
       fltp = (MYFLT *) p->auxch.auxp;
       p->rrev1 = fltp;   fltp += csound->ksmps;
